+++ src/app/cash/cash-routing.module.ts
@@ -1,0 +1,17 @@
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
+import { CashViewComponent } from './views';
+
+const routes: Routes = [
+  {
+    path: '',
+    pathMatch: 'full',
+    component: CashViewComponent,
+  },
+];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule],
+})
+export class CashRoutingModule {}