+++ src/app/cash/components/index.ts
@@ -1,0 +1,4 @@
+export * from './cash-transaction-dialog/cash-transaction-dialog.component';
+export * from './forex-transaction-dialog/forex-transaction-dialog.component';
+export * from './stock-transaction-dialog/stock-transaction-dialog.component';
+export * from './transaction-table/transaction-table.component';