+++ src/app/shared/types/index.ts
@@ -2,3 +2,10 @@ export type ErrorResponse = {
   error: Record<string, string[] | string>;
   status: number;
 };
+
+export type ApiPaginationResponse<T = unknown> = {
+  count?: number;
+  results: T[];
+};
+
+export type Currency = 'USD' | 'EUR' | 'HUF';