+++ src/app/stocks/components/index.ts
@@ -2,4 +2,3 @@ export * from './basic-stock-data-table/basic-stock-data-table.component';
 export * from './stock-dividend-table/stock-dividend-table.component';
 export * from './stock-pnl-table/stock-pnl-table.component';
 export * from './stock-risk-table/stock-risk-table.component';
-export * from './stock-transaction-dialog/stock-transaction-dialog.component';