+++ src/app/stocks/interfaces/index.ts
@@ -1,5 +1,5 @@
 export * from './stock-portfolio-summary.interface';
 export * from './stock-portfolio.interface';
 export * from './stock-positions.interface';
-export * from './stock-transaction-interface';
+export * from './stock-transaction.interface';
 export * from './stock.interface';