+++ src/app/stocks/services/stocks.service.ts
@@ -1,8 +1,9 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { finalize, switchMap, tap } from 'rxjs/operators';
+import { finalize, pluck, tap } from 'rxjs/operators';
 import { defaultCatchError } from 'src/app/shared/operators';
-import { Stock, StockTransaction } from '../interfaces';
+import { ApiPaginationResponse } from 'src/app/shared/types';
+import { Stock } from '../interfaces';
 import { StockPortfolioQuery, StockStore } from '../state';
 import { PortfolioService } from './portfolio.service';
 
@@ -21,26 +22,14 @@ export class StockService {
     this.store.setLoading(true);
 
     this.http
-      .get<Stock[]>(`/stocks/`)
+      .get<ApiPaginationResponse<Stock>>(`/stocks/`)
       .pipe(
+        pluck('results'),
         tap((stocks) => {
           this.store.set(stocks);
         }),
         defaultCatchError(this.store),
         finalize(() => this.store.setLoading(false))
-      )
-      .subscribe();
-  }
-
-  public createTransaction(transaction: StockTransaction): void {
-    this.store.setLoading(true);
-
-    this.http
-      .post<StockTransaction>(`/transactions/stocks`, transaction)
-      .pipe(
-        switchMap(() => this.portfolioQuery.selectedPortfolio),
-        tap((portfolio) => this.portfolioService.summary(portfolio?.id)),
-        defaultCatchError(this.store)
       )
       .subscribe();
   }